+++ pypolychord/settings.py
@@ -160,6 +160,17 @@ class PolyChordSettings:
         Note **Positive seeds only**
         a negative seed indicates that you should use the system time in
         milliseconds
+    prior_samples: array-like
+        (Default: None)
+        samples from the unit hypercube to start nested sampling from. This is
+        useful if the prior is hard to rejection sample directly from the unit
+        hypercube (e.g. if there is a large amount of excluded parameter
+        space), but can be done more efficiently manually. A concrete example
+        could be include sorted uniform parameters by requiring 0 < x1 < x2 <
+        x3 < 1, if one did not want to use SortedUniformPrior.
+        Only available in Python interface.
+        shape (:, nDims)
+
 
     """
     def __init__(self, nDims, nDerived, **kwargs):
@@ -196,6 +207,7 @@ class PolyChordSettings:
         self.grade_frac = list(kwargs.pop('grade_frac',
                                           [1.0]*len(self.grade_dims)))
         self.nlives = kwargs.pop('nlives', {})
+        self.prior_samples = kwargs.pop('prior_samples', None)
 
         if kwargs:
             raise TypeError('Unexpected **kwargs in Contours constructor: %r'
