+++ requirements.txt
@@ -1,3 +1,3 @@
 numpy
 scipy
-anesthetic==2.0.0b11
+anesthetic>=2.0.0b11
